Extract @nteract/myths package
