+++ packages/epics/tsconfig.json
@@ -9,6 +9,7 @@
     { "path": "../actions" },
     { "path": "../commutable" },
     { "path": "../messaging" },
+    { "path": "../mythic-notifications" },
     { "path": "../selectors" },
     { "path": "../types" },
     { "path": "../rx-jupyter" }
