+++ packages/mythic-notifications/src/myths/initialize-system.tsx
@@ -1,50 +1,55 @@
 import { Classes, Toaster } from "@blueprintjs/core";
-import { selectors } from "@nteract/core";
+import { MythicComponent } from "@nteract/myths";
+import { AppState } from "@nteract/types";
 import React, { RefObject } from "react";
 import styled from "styled-components";
 import { blueprintjsNotificationSystem } from "../backends/blueprintjs";
-import { MythicComponent } from "../external/myths";
 import { notifications } from "../package";
+import { NotificationSystem } from "../types";
 
 const initializeSystem =
-  notifications.createMyth("initializeSystem")({
+  notifications.createMyth("initializeSystem")<NotificationSystem>({
     reduce: (state, action) =>
-      state.set("current", action.payload.current),
+      state.set("current", action.payload),
   });
 
-const DoNotPrint = styled.div`
-  @media print {
-    display: none;
-  }
-`;
-
-export const NotificationRoot = initializeSystem.createConnectedComponent(
-  "NotificationRoot",
-  class extends MythicComponent<typeof initializeSystem, { theme: string }> {
-    toaster?: RefObject<Toaster>;
-
-    postConstructor(): void {
-      this.toaster = React.createRef();
+const DoNotPrint =
+  styled.div`
+    @media print {
+      display: none;
     }
+  `;
 
-    componentDidMount(): void {
-      this.props.initializeSystem({
-        current: blueprintjsNotificationSystem(this.toaster!.current!),
-      });
-    }
+export const NotificationRoot =
+  initializeSystem.createConnectedComponent(
+    "NotificationRoot",
+    class extends MythicComponent<typeof initializeSystem, { theme: string }> {
+      toaster?: RefObject<Toaster>;
 
-    render(): JSX.Element {
-      return (
-        <DoNotPrint>
-          <Toaster
-            ref={this.toaster}
-            usePortal={false}
-            position={"top-right"}
-            className={this.props.theme === "dark" ? Classes.DARK : undefined}
-          />
-        </DoNotPrint>
-      );
-    }
-  },
-  (state) => ({theme: selectors.currentTheme(state)}),
-);
+      postConstructor(): void {
+        this.toaster = React.createRef();
+      }
+
+      componentDidMount(): void {
+        this.props.initializeSystem(
+          blueprintjsNotificationSystem(this.toaster!.current!)
+        );
+      }
+
+      render(): JSX.Element {
+        return (
+          <DoNotPrint>
+            <Toaster
+              ref={this.toaster}
+              position={"top-right"}
+              className={this.props.theme === "dark" ? Classes.DARK : undefined}
+              usePortal={false}   // needed for the theme class to bubble down
+            />
+          </DoNotPrint>
+        );
+      }
+    },
+    (state: AppState) => ({
+      theme: state.config.get("theme"),
+    }),
+  );
