+++ packages/mythic-notifications/src/package.ts
@@ -1,5 +1,5 @@
+import { createMythicPackage } from "@nteract/myths";
 import { consoleNotificationSystem } from "./backends/console";
-import { createMythicPackage } from "./external/myths";
 import { NotificationSystem } from "./types";
 
 export const notifications = createMythicPackage("notifications")<
