+++ packages/myths/src/index.ts
@@ -1,0 +1,3 @@
+export { createMythicPackage } from "./package";
+export { MythicComponent } from "./react";
+export * from "./types";
