+++ packages/myths/tsconfig.json
@@ -4,9 +4,5 @@
     "outDir": "lib",
     "rootDir": "src"
   },
-  "include": ["src"],
-  "references": [
-    { "path": "../myths" },
-    { "path": "../types" }
-  ]
+  "include": ["src"]
 }
