+++ packages/tsconfig.json
@@ -22,6 +22,7 @@
     { "path": "messaging" },
     { "path": "monaco-editor" },
     { "path": "mythic-notifications" },
+    { "path": "myths" },
     { "path": "notebook-app-component" },
     { "path": "octicons" },
     { "path": "outputs" },
